BOILERPLATE/BOILERPLATE_MONGO/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# MongoDB
motor>=3.3.2  # Async MongoDB driver
pymongo>=4.6.1  # MongoDB driver (motor dependency)

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Validation and settings
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP client
httpx>=0.26.0

# Development
python-dotenv>=1.0.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Optional: For migrations/seeding
# mongoengine>=0.27.0
